+++ imp/src/runtime/api/Api.org~
@@ -178,63 +178,10 @@ import Data.Text
 import Database.Persist
 
 import Models
+import SecureApi
 import Servant.API
 #+END_SRC
 
-* Custom Datatypes For Instruction Input
-  
-Defining custom datatypes for the request body when required.
-
-** =UniqueUserData= : a unique value to identify the user by
-
-In this case, we take =UniqueUserData= to be =userEmail=.
-
-#+NAME: unique_user_data
-#+BEGIN_SRC haskell
-
-data UniqueUserData = UniqueUserData {
-                                       userData :: String
-                                     } deriving (Eq, Read, Show)
-
-
-instance ToJSON UniqueUserData where
-  toJSON (UniqueUserData userData) =
-    object ["data" .= userData]
-
-instance FromJSON UniqueUserData where
-  parseJSON = withObject "UniqueUserData" $ \ v ->
-    UniqueUserData <$> v .: "data"
-
-
-
-#+END_SRC
-
-* Custom Datatypes for Response
-
-Defining custom datatypes/instances for response.
-
-** ResponseUserId
-
-#+NAME: response_user_id
-#+BEGIN_SRC haskell
-data ResponseUserId = ResponseUserId {
-                                     userIdValue :: (Key (User))
-                                     } deriving (Generic, Eq, Read, Show)
-
-
-                      
-
-#+END_SRC
-** ResponseSessionId
-
-#+NAME: response_session_id
-#+BEGIN_SRC haskell
-data ResponseSessionId = ResponseSessionId {
-                                     sessionIdValue :: (Key (Session))
-                                     } deriving (Generic, Eq, Read, Show)
-
-
-#+END_SRC
 * API 
 
 ** Routes not requiring authentication
@@ -268,10 +215,6 @@ type UserAPI = NonSecureRoutes
 
 #+BEGIN_SRC haskell :eval no :noweb yes :tangle Api.hs
 <<extns_and_imports>>
-<<unique_user_data>>
-<<response_user_id>>
-<<response_session_id>>
 <<nonsecure_routes>>
-<<secure_routes>>
 <<combined_api>>
 #+END_SRC
