+++ imp/src/runtime/api/NonSecureRoutes.org
@@ -1,0 +1,71 @@
+#+TITLE: Api Routes Not Protected By Auth
+
+
+* Introduction 
+
+Defining the portion of the Api that do not require login to access.
+* Language Extensions and Imports 
+
+#+NAME: extns_and_imports
+#+BEGIN_SRC haskell 
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeOperators #-}
+
+module NonSecureApi where
+
+import Prelude ()
+import Prelude.Compat
+
+import Control.Monad.Except
+import Control.Monad.Reader
+import Data.Aeson.Compat
+import Data.Aeson.Types
+import Data.Attoparsec.ByteString
+import Data.ByteString (ByteString)
+import Data.List
+import Data.Maybe
+import Data.String.Conversions
+import Data.Time.Calendar
+import GHC.Generics
+import Lucid
+import Network.HTTP.Media ((//), (/:))
+import Network.Wai
+import Network.Wai.Handler.Warp
+import Servant
+import System.Directory
+import Text.Blaze
+import Text.Blaze.Html.Renderer.Utf8
+import qualified Data.Aeson.Parser
+import qualified Text.Blaze.Html
+
+import Data.Proxy
+import Data.Text
+
+import Database.Persist
+
+import Models
+
+import Servant.API
+#+END_SRC
+* API
+
+#+NAME: non_secure_routes
+#+BEGIN_SRC haskell
+
+type NonSecureRoutes = "index" :> Get '[PlainText] Text
+                  :<|> "login" :> ReqBody '[JSON] Session :> Post '[JSON] (Maybe (ResponseSessionId))
+
+#+END_SRC
+* Tangling
+
+#+BEGIN_SRC haskell :eval no :noweb yes :tangle NonSecureApi.hs 
+<<extns_and_imports>>
+<<non_secure_routes>>
+
+#+END_SRC
