+++ imp/src/runtime/api/SecureApi.org~
@@ -1,0 +1,74 @@
+#+TITLE: Secure Routes For User Dir Api
+
+
+* Language Extensions and Imports 
+
+#+NAME: extns_and_imports
+#+BEGIN_SRC haskell 
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeOperators #-}
+
+module SecureApi where
+
+import Prelude ()
+import Prelude.Compat
+
+import Control.Monad.Except
+import Control.Monad.Reader
+import Data.Aeson.Compat
+import Data.Aeson.Types
+import Data.Attoparsec.ByteString
+import Data.ByteString (ByteString)
+import Data.List
+import Data.Maybe
+import Data.String.Conversions
+import Data.Time.Calendar
+import GHC.Generics
+import Lucid
+import Network.HTTP.Media ((//), (/:))
+import Network.Wai
+import Network.Wai.Handler.Warp
+import Servant
+import System.Directory
+import Text.Blaze
+import Text.Blaze.Html.Renderer.Utf8
+import qualified Data.Aeson.Parser
+import qualified Text.Blaze.Html
+
+import Data.Proxy
+import Data.Text
+
+import Database.Persist
+
+import Models
+import Servant.API
+#+END_SRC
+
+* Routes
+
+#+NAME: secure_routes
+#+BEGIN_SRC haskell 
+
+ 
+type SecureRoutes = Header "Cookie" String :> "showUsers" :> Get '[JSON] [User]
+          :<|> Header "Cookie" String :> "addUser" :> ReqBody '[JSON] User :> Post '[JSON] (Maybe (ResponseUserId))
+          :<|> Header "Cookie" String :> "deleteUser" :> ReqBody '[JSON] UniqueUserData :> Post '[JSON] (Maybe (User))
+          :<|> Header "Cookie" String :> "logout" :> ReqBody '[JSON] Session :> Post '[JSON] (Maybe (Session))
+
+
+
+
+#+END_SRC
+* Tangling
+
+#+BEGIN_SRC haskell :eval no :noweb yes :tangle SecureApi.hs
+<<extns_and_imports>>
+<<secure_routes>>
+
+#+END_SRC
