+++ imp/src/runtime/controllers/Controllers.org
@@ -1,0 +1,143 @@
+#+TITLE: Controllers For User Directory Webapp
+
+
+* Introduction
+
+Defining the controllers to perform CRUD operations.
+* Language Extensions and Imports
+
+  
+#+NAME: extns_and_imports
+#+BEGIN_SRC haskell
+{-# LANGUAGE DataKinds         #-}
+{-# LANGUAGE DeriveGeneric     #-}
+{-# LANGUAGE LambdaCase        #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeFamilies      #-}
+{-# LANGUAGE TypeOperators     #-}
+
+module Controllers where
+
+import           Data.Aeson
+import           Data.Text
+import           Control.Monad.IO.Class
+import           Control.Monad.Logger (runStderrLoggingT)
+
+import           Data.String.Conversions
+import qualified Data.ByteString.Lazy.Char8 as B           
+
+import           Database.Persist
+import           Database.Persist.Sql
+import           Database.Persist.Sqlite
+
+import           Network.Wai
+import           Network.Wai.Handler.Warp as Warp
+
+import           Servant
+
+import           Data.Text
+
+import           Api
+import           Models
+import           Role
+import           Authentication 
+
+#+END_SRC
+* Controllers
+
+** showUsers controller
+
+#+NAME: show_all_users_helper
+#+BEGIN_SRC haskell 
+
+-- helper function for showUsersHandler
+showAllUsersHelper :: ConnectionPool -> Bool -> IO ([User])
+showAllUsersHelper pool authVal = flip runSqlPersistMPool pool $ case authVal of
+    False -> return []
+    True  -> do
+      users <- selectList [] []
+      return $ Prelude.map entityVal users
+
+#+END_SRC
+
+** addUser controller
+
+#+NAME: add_user_helper
+#+BEGIN_SRC haskell
+
+-- helper function for addUserHandler
+addUserHelper :: User -> ConnectionPool -> Bool -> IO (Maybe (ResponseUserId))
+addUserHelper newUser pool authVal = flip runSqlPersistMPool pool $ case authVal of
+  False -> return Nothing
+  True  -> do
+    exists <- selectFirst [UserName ==. (userName newUser)] []
+    case exists of
+      Nothing -> Just <$> (ResponseUserId <$> insert newUser)
+      Just _  -> return Nothing
+
+#+END_SRC
+
+** deleteUser controller
+
+#+NAME: delete_user_helper
+#+BEGIN_SRC haskell
+
+-- helper function for deleteUserHandler
+deleteUserHelper :: Text -> ConnectionPool -> Bool -> IO ((Maybe (User)))
+deleteUserHelper userToDel pool authVal = flip runSqlPersistMPool pool $ case authVal of
+  False -> return Nothing
+  True  -> do
+    deletedUser <- selectFirst [UserEmail ==. unpack(userToDel)] []
+    case deletedUser of
+      Nothing -> return Nothing
+      Just _ -> do
+        userIfDeleted <- deleteWhere [UserEmail ==. unpack(userToDel)]
+        return $ entityVal <$> deletedUser
+
+#+END_SRC
+
+** login controller
+
+#+NAME: login_helper
+#+BEGIN_SRC haskell
+
+-- helper function for loginHandler
+loginHelper :: Session -> ConnectionPool -> IO (Maybe (ResponseSessionId))
+loginHelper newSession pool = flip runSqlPersistMPool pool $ do
+  ifExists <- selectFirst [UserEmail ==. (sessionUserEmail newSession)] []
+  case ifExists of
+    Nothing -> return Nothing
+    Just _  -> Just <$> (ResponseSessionId <$> insert newSession)
+  
+#+END_SRC
+
+** logout controller
+
+#+NAME: logout_helper
+#+BEGIN_SRC haskell
+
+-- helper function for logoutHandler
+logoutHelper :: Session -> ConnectionPool -> Bool -> IO (Maybe (Session))
+logoutHelper currentSession pool authVal = flip runSqlPersistMPool pool $ case authVal of
+  False -> return Nothing
+  True  -> do
+    ifExists <- selectFirst [SessionUserEmail ==. (sessionUserEmail currentSession), SessionUserRoles ==. (sessionUserRoles currentSession)] []
+    case ifExists of
+      Nothing -> return Nothing
+      Just _ -> do
+        deleteWhere [SessionUserEmail ==. (sessionUserEmail currentSession)]
+        return $ entityVal <$> ifExists
+      
+  
+#+END_SRC
+
+* Tangling
+
+#+BEGIN_SRC haskell :eval no :noweb yes :tangle Controllers.hs
+<<extns_and_imports>>
+<<show_all_users_helper>>
+<<add_user_helper>>
+<<delete_user_helper>>
+<<login_helper>>
+<<logout_helper>>
+#+END_SRC
