+++ imp/src/runtime/models/models.org~
@@ -39,6 +39,7 @@ import           Database.Persist
 import           Database.Persist.Sqlite
 import           Database.Persist.TH
 
+import           GHC.Generics
 #+END_SRC
 * Externally Defined Datatypes
 
@@ -105,10 +106,67 @@ Session
 
 #+END_SRC
 
+* Custom Datatypes For Instruction Input
+  
+Defining custom datatypes for the request body when required.
+
+** =UniqueUserData= : a unique value to identify the user by
+
+In this case, we take =UniqueUserData= to be =userEmail=.
+
+#+NAME: unique_user_data
+#+BEGIN_SRC haskell
+
+data UniqueUserData = UniqueUserData {
+                                       userData :: String
+                                     } deriving (Eq, Read, Show)
+
+
+instance ToJSON UniqueUserData where
+  toJSON (UniqueUserData userData) =
+    object ["data" .= userData]
+
+instance FromJSON UniqueUserData where
+  parseJSON = withObject "UniqueUserData" $ \ v ->
+    UniqueUserData <$> v .: "data"
+
+
+
+#+END_SRC
+
+* Custom Datatypes for Response
+
+Defining custom datatypes/instances for response.
+
+** ResponseUserId
+
+#+NAME: response_user_id
+#+BEGIN_SRC haskell
+data ResponseUserId = ResponseUserId {
+                                     userIdValue :: (Key (User))
+                                     } deriving (Eq, Read, Show)
+
+
+                      
+
+#+END_SRC
+** ResponseSessionId
+
+#+NAME: response_session_id
+#+BEGIN_SRC haskell
+data ResponseSessionId = ResponseSessionId {
+                                     sessionIdValue :: (Key (Session))
+                                     } deriving (Eq, Read, Show)
+
+
+#+END_SRC
 * Tangling 
 
 #+NAME: tangling
 #+BEGIN_SRC haskell :eval no :noweb yes :tangle Models.hs
 <<extns_and_imports>>
 <<database_models>>
+<<unique_user_data>>
+<<response_user_id>>
+<<response_session_id>>
 #+END_SRC
