+++ imp/src/test/AppSpec.org
@@ -112,7 +112,7 @@ userAPI = Data.Proxy.Proxy
 
 
 -- code that returns the client functions for our API
-(testIndex :<|> testLogin) :<|> (testShowUsers :<|> testAddUser :<|> testDeleteUser :<|> testLogout) = client userAPI 
+(testIndex :<|> testLogin) :<|> (testLogout) :<|> (testShowUsers :<|> testAddUser :<|> testDeleteUser) = client userAPI 
 
 #+END_SRC
